+++ web/src/main.tsx
@@ -1,10 +1,16 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App.js";
+import { initAnalytics } from "./analytics.js";
+import { AppErrorBoundary } from "./components/AppErrorBoundary.js";
 import "./index.css";
+
+initAnalytics();
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    <App />
+    <AppErrorBoundary>
+      <App />
+    </AppErrorBoundary>
   </StrictMode>
 );